Rename servlet-javax-common module and packages to follow new convention for common modules
