+++ instrumentation/servlet/servlet-2.2/javaagent/build.gradle.kts
@@ -19,7 +19,7 @@ muzzle {
 
 dependencies {
   implementation(project(":instrumentation:servlet:servlet-common:javaagent"))
-  implementation(project(":instrumentation:servlet:servlet-javax-common:library"))
+  implementation(project(":instrumentation:servlet:servlet-common-javax:library"))
   bootstrap(project(":instrumentation:servlet:servlet-common:bootstrap"))
 
   compileOnly("javax.servlet:servlet-api:2.2")
