+++ instrumentation/servlet/servlet-common-javax/library/src/main/java/io/opentelemetry/instrumentation/servlet/common/javax/JavaxServletAccessor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.servlet.javax;
+package io.opentelemetry.instrumentation.servlet.common.javax;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
