+++ settings.gradle.kts
@@ -614,7 +614,7 @@ include(":instrumentation:servlet:servlet-5.0:tomcat-testing")
 include(":instrumentation:servlet:servlet-common:bootstrap")
 include(":instrumentation:servlet:servlet-common:javaagent")
 include(":instrumentation:servlet:servlet-common:library")
-include(":instrumentation:servlet:servlet-javax-common:library")
+include(":instrumentation:servlet:servlet-common-javax:library")
 include(":instrumentation:spark-2.3:javaagent")
 include(":instrumentation:spring:spring-batch-3.0:javaagent")
 include(":instrumentation:spring:spring-boot-actuator-autoconfigure-2.0:javaagent")
